HW_16_Flask/requirements.txt
Flask==2.2.2
SQLAlchemy
python-dotenv
pymysql
werkzeug.security